✨ [#10] 주식 게임 라운드 진행 구현

+++ src/main/java/com/example/demo/stock/entity/Year.java
@@ -18,7 +18,7 @@ public class Year {
 
     @Id
     @Column(name = "year_id", nullable = false)
-    private Integer yearId;
+    private Long yearId;
 
     @Column(name = "hint1")
     private String hint1;
